Moved StreetName related values to separate dir
