+++ src/Normalizer/FromJson/StreetNameDetailNormalizer.php
@@ -4,8 +4,8 @@ declare(strict_types=1);
 
 namespace DigipolisGent\Flanders\BasicRegisters\Normalizer\FromJson;
 
-use DigipolisGent\Flanders\BasicRegisters\Value\StreetNameDetail;
-use DigipolisGent\Flanders\BasicRegisters\Value\StreetNameId;
+use DigipolisGent\Flanders\BasicRegisters\Value\Street\StreetNameDetail;
+use DigipolisGent\Flanders\BasicRegisters\Value\Street\StreetNameId;
 
 /**
  * Normalizes JSON data into a StreetNameDetail value.
@@ -17,7 +17,7 @@ final class StreetNameDetailNormalizer
      *
      * @param object $jsonData
      *
-     * @return \DigipolisGent\Flanders\BasicRegisters\Value\StreetNameDetail
+     * @return \DigipolisGent\Flanders\BasicRegisters\Value\Street\StreetNameDetail
      */
     public function normalize(object $jsonData): StreetNameDetail
     {
