+++ src/Normalizer/FromJson/StreetNameNormalizer.php
@@ -4,8 +4,8 @@ declare(strict_types=1);
 
 namespace DigipolisGent\Flanders\BasicRegisters\Normalizer\FromJson;
 
-use DigipolisGent\Flanders\BasicRegisters\Value\StreetName;
-use DigipolisGent\Flanders\BasicRegisters\Value\StreetNameId;
+use DigipolisGent\Flanders\BasicRegisters\Value\Street\StreetName;
+use DigipolisGent\Flanders\BasicRegisters\Value\Street\StreetNameId;
 
 /**
  * Normalizes JSON data into a StreetName object.
@@ -17,7 +17,7 @@ final class StreetNameNormalizer
      *
      * @param object $jsonData
      *
-     * @return \DigipolisGent\Flanders\BasicRegisters\Value\StreetName
+     * @return \DigipolisGent\Flanders\BasicRegisters\Value\Street\StreetName
      */
     public function normalize(object $jsonData): StreetName
     {
