+++ src/Value/Street/StreetName.php
@@ -2,8 +2,10 @@
 
 declare(strict_types=1);
 
-namespace DigipolisGent\Flanders\BasicRegisters\Value;
+namespace DigipolisGent\Flanders\BasicRegisters\Value\Street;
 
+use DigipolisGent\Flanders\BasicRegisters\Value\AbstractWithGeographicalName;
+use DigipolisGent\Flanders\BasicRegisters\Value\GeographicalName;
 use DigipolisGent\Value\ValueInterface;
 
 /**
@@ -14,14 +16,14 @@ final class StreetName extends AbstractWithGeographicalName
     /**
      * The street name id.
      *
-     * @var \DigipolisGent\Flanders\BasicRegisters\Value\StreetNameId
+     * @var \DigipolisGent\Flanders\BasicRegisters\Value\Street\StreetNameId
      */
     private $streetNameId;
 
     /**
      * Create a new street name.
      *
-     * @param \DigipolisGent\Flanders\BasicRegisters\Value\StreetNameId $streetNameNameId
+     * @param \DigipolisGent\Flanders\BasicRegisters\Value\Street\StreetNameId $streetNameNameId
      * @param \DigipolisGent\Flanders\BasicRegisters\Value\GeographicalName $geographicalName
      */
     public function __construct(StreetNameId $streetNameNameId, GeographicalName $geographicalName)
@@ -33,7 +35,7 @@ final class StreetName extends AbstractWithGeographicalName
     /**
      * Get the street name id.
      *
-     * @return \DigipolisGent\Flanders\BasicRegisters\Value\StreetNameId
+     * @return \DigipolisGent\Flanders\BasicRegisters\Value\Street\StreetNameId
      */
     public function streetNameId(): StreetNameId
     {
@@ -45,7 +47,7 @@ final class StreetName extends AbstractWithGeographicalName
      */
     public function sameValueAs(ValueInterface $object): bool
     {
-        /** @var \DigipolisGent\Flanders\BasicRegisters\Value\StreetName $object */
+        /** @var \DigipolisGent\Flanders\BasicRegisters\Value\Street\StreetName $object */
         return parent::sameValueAs($object)
             && $this->streetNameId()->sameValueAs($object->streetNameId());
     }
