+++ src/Value/Street/StreetNameId.php
@@ -1,0 +1,15 @@
+<?php
+
+declare(strict_types=1);
+
+namespace DigipolisGent\Flanders\BasicRegisters\Value\Street;
+
+use DigipolisGent\Flanders\BasicRegisters\Value\AbstractId;
+
+/**
+ * Street name object ID.
+ */
+final class StreetNameId extends AbstractId
+{
+
+}
