--- src/Value/StreetNameId.php
@@ -1,13 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace DigipolisGent\Flanders\BasicRegisters\Value;
-
-/**
- * Street name object ID.
- */
-final class StreetNameId extends AbstractId
-{
-
-}
