+++ tests/Normalize/FromJson/StreetNamesNormalizerTest.php
@@ -7,9 +7,9 @@ namespace DigipolisGent\Tests\Flanders\BasicRegisters\Normalize\FromJson;
 use DigipolisGent\Flanders\BasicRegisters\Normalizer\FromJson\StreetNamesNormalizer;
 use DigipolisGent\Flanders\BasicRegisters\Value\GeographicalName;
 use DigipolisGent\Flanders\BasicRegisters\Value\LanguageCode;
-use DigipolisGent\Flanders\BasicRegisters\Value\StreetName;
-use DigipolisGent\Flanders\BasicRegisters\Value\StreetNameId;
-use DigipolisGent\Flanders\BasicRegisters\Value\StreetNames;
+use DigipolisGent\Flanders\BasicRegisters\Value\Street\StreetName;
+use DigipolisGent\Flanders\BasicRegisters\Value\Street\StreetNameId;
+use DigipolisGent\Flanders\BasicRegisters\Value\Street\StreetNames;
 use PHPUnit\Framework\TestCase;
 
 /**
