+++ tests/Value/Street/StreetNameDetailTest.php
@@ -2,20 +2,20 @@
 
 declare(strict_types=1);
 
-namespace DigipolisGent\Tests\Flanders\BasicRegisters\Value;
+namespace DigipolisGent\Tests\Flanders\BasicRegisters\Value\Street;
 
 use DigipolisGent\Flanders\BasicRegisters\Value\GeographicalName;
 use DigipolisGent\Flanders\BasicRegisters\Value\GeographicalNames;
 use DigipolisGent\Flanders\BasicRegisters\Value\LanguageCode;
 use DigipolisGent\Flanders\BasicRegisters\Value\Locality\LocalityName;
 use DigipolisGent\Flanders\BasicRegisters\Value\Locality\LocalityNameId;
-use DigipolisGent\Flanders\BasicRegisters\Value\StreetNameDetail;
-use DigipolisGent\Flanders\BasicRegisters\Value\StreetNameId;
+use DigipolisGent\Flanders\BasicRegisters\Value\Street\StreetNameDetail;
+use DigipolisGent\Flanders\BasicRegisters\Value\Street\StreetNameId;
 use PHPUnit\Framework\TestCase;
 
 /**
  * @covers \DigipolisGent\Flanders\BasicRegisters\Value\AbstractWithGeographicalNames
- * @covers \DigipolisGent\Flanders\BasicRegisters\Value\StreetNameDetail
+ * @covers \DigipolisGent\Flanders\BasicRegisters\Value\Street\StreetNameDetail
  */
 class StreetNameDetailTest extends TestCase
 {
